backend/requirements.txt
asgiref==3.6.0
astroid==2.15.0
black==23.1.0
click==8.1.3
dill==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
isort==5.12.0
lazy-object-proxy==1.9.0
loguru==0.6.0
Markdown==3.4.1
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
platformdirs==3.1.0
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pylint==2.16.4
pytz==2022.7.1
sqlparse==0.4.3
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
wrapt==1.15.0
